refactor: improve affected packages detection and cleanup scripts
- Update affected.ts to use new directory detection utility
- Enhance local-cleanup.ts with better error handling
- Add get-all-directories.ts utility for consistent directory scanning
- Improve script reliability and maintainability
